Refactor Predictions tab into separate (presentational and container) component files
