+++ static/js/containers/PredictionsTab/index.jsx
@@ -1,0 +1,14 @@
+import { connect } from 'react-redux';
+
+import * as Action from '../../actions';
+import PredictTab from '../../components/PredictionsTab';
+
+
+const mapDispatchToProps = dispatch => (
+  {
+    doPrediction: form => dispatch(Action.doPrediction(form))
+  }
+);
+
+
+export default connect(null, mapDispatchToProps)(PredictTab);
